Remove unnecessary package overlay declaration from debug manifest
I modelled (nice word for copied) this screenshots automation setup from
what the WooCommerce Android app does. Because it's debug manifest did declare
the package, see
https://github.com/woocommerce/woocommerce-android/blob/develop/WooCommerce/src/debug/AndroidManifest.xml#L3
I made this one do too.

I did notice, though, that this app has `.debug` appended to its debug
package, so that's what I used in the manifest. That gave me an error:

```
Manifest merger failed :

Overlay manifest:package attribute declared at AndroidManifest.xml:3:11-52
value=(com.automattic.simplenote.debug) has a different
value=(com.automattic.simplenote.debug) declared in main manifest at
AndroidManifest.xml:4:5-40

Suggestion: remove the overlay declaration at AndroidManifest.xml and place it in
the build.gradle:

flavorName {
        applicationId = "com.automattic.simplenote.debug"
}
```

I gave it a go the same way Woo does, and it worked.

During code review,
https://github.com/Automattic/simplenote-android/pull/1026#discussion_r423721885,
@theck13 raised the question of whether the package
was appropriate, which prompted me to have a better read of the error
and realize it suggests to "remove the overlay declaration", which is
what this commit does.

I run it both from Android Studio and from Fastlane and it worked fine.

I'm guessing that between the settings in the build.gradle and the
convention of where the manifest is located in the file system, the
build understands that we are dealing with the debug version even though
the package is not explicit.