Use a longer timeout after login for the screenshots automation
While trying to reproduce the failure from this comment,
https://github.com/Automattic/simplenote-android/pull/1038#issuecomment-635717561,
I occasionally got a timeout when waiting for the target note to be on
screen after a login. The fact is, the target not was already on screen.
I think this kind of failure is consistent with my supposition that
there's something getting lost when the code use the root view to
perform the wait.

I hope that by using a longer timeout, the root view will be the one we
expect consistently.