Fix changing device language crash with a note open
The crash was caused due to the fact that, when saving a tab during
'onSaveInstanceState(...)' a key of a specific language was used
(ie. 'Edit' if the device was on English at that point), but when
restoring a tab during 'onCreate(...)' a key of another language was
used (ie. 'Editar' if the device was switched to Spanish). Thus, no
restore actually occurred on the tab layer, causing the crash.

Using a constant as a key for saving/restoring state fixes this issue.