Note Info Sheet: Update with updates to note
Previously the note info sheet would render the information about a note on open
and then statically hold those values until it's closed and opened again.

In this patch we are adding listenwers to the application to watch for changes
to the note and then re-displaying the info so that it stays current.

There are a couple caveats:
 - Because I wasn't sure how the UI is being arranged the references don't update.
   When I blindly called `getReferences(note)` on updates it blanked out the
   list of references.
 - Similarly I tried setting the visibility of the sync label to `VISIBLE` when
   updating the sync value but the label doesn't appear until re-opening the sheet.
   I think this is related to the references display issue.

Therefore in order to see the sync time if the sheet was opened when none was available
you need to re-open the sheet. If you open the sheet and a sync time is already avialable
it will work as expected. If you want to see the updated references that reflect changes
in other notes you will also need to re-open the sheet.