Extract delay fetch for NoteMarkdownFragment to before page finishes
While working on the screenshot automation, the tests started
consistently failing after dismissing a note editor activity because the
NoteMarkdownFragment was not attached to a context from the call edited
in this commit.

I'm not sure why that would happen or if its of any concern in
production.

With my limited Android knowledge, this seems to be a legit workaround.
We access the context at a time in which we are sure it exists. The
value we extract is constant over the application lifecycle anyways, so
there's no downside in reading it from the context at a different time
than when it's going to be used.