Remove tags in notes comparing them with the canonical representation
The notes keep a list of tags associated. The tags saved are the name
of the tags and not their canonical representation. To remove the
correct tags, we need to compare the canonical representation of the
tag to remove against that canonical representation of the tags
associated with the notes.