Separate setup initial UiState and start listening to tag changes in TagsViewModel
The previous implementation of the method start() setup the initial UiState with all tags and empty search term. It also starts listening to tag changes from the tag's bucket. The start() method was called in onResume, thus if the user leave the app and comeback (the activity is restarted), the start() method was called everytime, cleaning the UiState to its initial state. This created a problem when the user search tags and leave the app because the search term stayed in the search bar but it showed the complete list of tags.
This commit separates the setup of the initial UiState (method start()) and start listening to tag changes (method startListeningTagChanges()). The method start() is now called in the onCreate() callback to avoid cleaning the UiState when the user returns an activity (activity is restarted).