Added a custom progress dialog fragment to handle different themes
The ProgressDialogFragment class that was used for progress bar dialogs
(part of Simperium library and used in the login) does not support dark
mode. When the dark mode was activated, the progress bar showed in
white. We implemented a custom layout which is used with an AlertDialog
to implement the same behavior as ProgressDialog but with support to
dark mode and other themes.