Set layout of delete account preference in XML file
Instead of setting the layout of the delete account preference in code,
we set the layout in the XML file. Moving this to the layout keep the
fragment from doing something the should be done in the layout. We also
simplify the layout and added just one layout for light and dark mode
(preferences_red.xml). This layout uses a user-defined color that is set
to red_50 for light and red_20 for dark themes.