Added initial version of review account verifier
Introduced an initial version to check whether we need to show the
review account dialog or not. This helper uses a worker which runs on
the background waiting until the account bucket is synchronized. When
the account bucket is synchronized, it uses the Simplenote applicaiton
class to show the review account dialog.