Removed check to review account when an activity is created
This is the main reason for the race condition. Everytime an activity
is created, there is a check for the review account. The problem with
this approach is that it is triggered just because the activity
SimplenoteCredentialsActivity is created everytime the user login.
Activities such as NotesActivity are just restarted. We need to make the
check after the user login into account. Thus, we remove the check for
review account when activities are created. This commit also added the
field mCurrentActivity to keep track of the last activity on the
foreground.