Changed account verifier to react when the account bucket is indexed
Moved the account verifier to check when that the account bucket
has been synchronized. We do that by implementing the callback handler
OnNetworkChangeListener which reacts when there is a change in the
account bucket.

When a user login into their account, Simplenote starts indexing the
different buckets, including the account bucket. We setup a
OnNetworkChangeListener on the account bucket to listen when the bucket
is indexed. One of the parameters of the callback onNetworkChange called
is Bucket.ChangeType. If this parameter is INDEX, then we know that the
account bucket was indexed.