Fixed dismiss method for FullScreenDialogFragment
The previous implementation was calling to pop the last fragment from
the fragment manager. This is a fragile action since there is not
assurance that the last fragment added was FullScreenDialogFragment.
We changed the call to super.dismiss() since both closes the dialog
and remove it from the fragments attached to the current activity.