Avoided dismissing dialog if application is in the background
For FullScreenDialogFragment, if we try to dismiss the dialog while the
application is in the background, it would cause an
IllegalStateException. We avoid such case by checking if the fragment
is not in save state. This happens in the case we show the dialog to
review account and the user turn off the screen or uses another app.
While in the background we receive an account modification that says the
account was reviewed, the application will try to dismiss the dialog,
which would crash the app.