Move scroll setup from view tree updates after the note content refresh
The initial scroll position is set based on the last scroll position the
user was or on the first keyword found if entering from search results.
To set the scroll position, the app used a ViewTreeObserver with and
setup a OnGlobalFocusChangeListener callback listener. The problem with
this approach is that this callback gets called several times even after
the text has been rendered. Thus, the scroll position for a search was
overridden by the scroll position saved in preferences.

We now change the scroll setup to after the note contents are refreshed.
This happens when the note is loaded and set in the EditText. We post a
task on the root view to be sure that the scroll setup is done after the
text has been rendered.