Fixes problem when layout is not ready for configuration change
When the configuration is changed and the user is searching through
keywords, we set the scroll on the last keyword selected. When the
setScroll is executed in some cases the edit text's layout has not been
rendered, therefore is not ready (is null). We recursively call the
method to make sure next time will be ready. This method avoids infinite
loop when the fragment is not added.