Use a predraw listener to check when field's layout becomes avaiable
After a configuration change, the layout of a the text editor is null.
We need to wait for some UI events to happen in order for the layout to
become available. To check that status, we use OnPreDrawListener on the
ViewTreeObserver and call setScroll when the layout is not null.