Remove `flavor` option — Probably never used
I couldn't see it ever being used in the `Fastfile` but I might have not
searched deep enough.

I suppose we never noticed it because we used to pass options in
hashes. Only moving to named block arguments revealed it.

I'm quite confident in this change because assembleVanillaRelease
fails due to there being no Vanilla.